+++ networking/references.js
@@ -1,0 +1,5 @@
+const BD_REFERENCES = {
+  pacientes: '/pacientes',
+}
+
+module.exports = { BD_REFERENCES };
